+++ main.tf
@@ -1,3 +1,4 @@
+##### Autoscaling
 locals {
   launch_template         = coalesce(var.launch_template_name, aws_launch_template.main[0].name, var.external_launch_template_name)
   launch_template_version = coalesce(var.launch_template_version, aws_launch_template.main[0].latest_version, var.external_launch_template_version)
@@ -5,9 +6,19 @@ locals {
 ############################
 ### Cloudwatch resources
 ############################
+resource "aws_kms_key" "cloudwatch" {
+  count                   = var.install_cloudwatch_agent ? 1 : 0
+  description             = "${var.name} Log Group KMS key"
+  enable_key_rotation     = var.enable_key_rotation
+  policy                  = element(concat(data.aws_iam_policy_document.kms.*.json, [""]), 0)
+  deletion_window_in_days = var.key_deletion_window_in_days
+}
+
 resource "aws_cloudwatch_log_group" "main" {
-  count = var.enable_monitoring ? 1 : 0
-  name  = "/aws/asg/${var.name}"
+  count             = var.install_cloudwatch_agent ? 1 : 0
+  name              = "/aws/asg/${var.name}"
+  retention_in_days = var.retention_in_days
+  kms_key_id        = aws_kms_key.cloudwatch[0].arn
   tags = merge(
     {
       "Name"        = var.name
@@ -29,32 +40,86 @@ resource "aws_key_pair" "main" {
   public_key = try(tls_private_key.main[0].public_key_openssh, null)
 }
 
+## For downloading the keypair to local computer
+resource "null_resource" "local_save_ec2_keypair" {
+  count = var.create_key_pair ? 1 : 0
+  provisioner "local-exec" {
+    command = "echo '${tls_private_key.main[0].private_key_pem}' > ${path.module}/${aws_key_pair.main[0].id}.pem"
+  }
+}
+
 ############################
 ### IAM Resources
 ############################
 resource "aws_iam_instance_profile" "main" {
-  name = "${var.name}_iam_role"
-  path = var.iam_role_path
-  role = aws_iam_role.main.name
+  count = var.create_instance_profile ? 1 : 0
+  name  = "${var.name}-iam-role"
+  path  = var.iam_role_path
+  role  = aws_iam_role.main[0].name
 }
 
 resource "aws_iam_role" "main" {
-  description        = "${var.name} ASG Group IAM Role"
-  name               = "${var.name}.iam_role"
+  count              = var.create_instance_profile ? 1 : 0
+  description        = "${var.name} EC2 IAM Role"
+  name               = "${var.name}-iam-role"
   path               = var.iam_role_path
   assume_role_policy = data.aws_iam_policy_document.assume_role.json
 }
 
 resource "aws_iam_policy" "main" {
-  name        = "${var.name}.iam_pol"
-  description = "${var.name} ASG Group IAM policy"
+  count       = var.create_instance_profile ? 1 : 0
+  name        = "${var.name}-iam-policy"
+  description = "${var.name} EC2 IAM role policy"
   path        = var.iam_role_path
   policy      = data.aws_iam_policy_document.asg.json
 }
 
 resource "aws_iam_role_policy_attachment" "main" {
-  role       = aws_iam_role.main.name
-  policy_arn = aws_iam_policy.main.arn
+  count      = var.create_instance_profile ? 1 : 0
+  role       = aws_iam_role.main[0].name
+  policy_arn = aws_iam_policy.main[0].arn
+}
+
+### For adding custom permissions to the role created above
+resource "aws_iam_policy" "additional" {
+  count       = var.additional_role_policy_document != null && var.create_instance_profile ? 1 : 0
+  name        = "${var.name}-additional-policy"
+  description = "${var.name} additional IAM role policy"
+  path        = var.iam_role_path
+  policy      = var.additional_role_policy_document
+}
+
+resource "aws_iam_role_policy_attachment" "additional" {
+  count      = var.additional_role_policy_document != null && var.create_instance_profile ? 1 : 0
+  role       = aws_iam_role.main[0].name
+  policy_arn = aws_iam_policy.additional[0].arn
+}
+
+## Managed Policy to allow cloudwatch agent to write metrics to CloudWatch
+resource "aws_iam_role_policy_attachment" "cloudwatchagentserverpolicy" {
+  count      = var.create_instance_profile && var.install_cloudwatch_agent ? 1 : 0
+  policy_arn = "arn:aws:iam::aws:policy/CloudWatchAgentServerPolicy"
+  role       = aws_iam_role.main[0].name
+}
+
+## Configure CloudWatch agent to set the retention policy for log groups that it sends log events to.
+resource "aws_iam_role_policy" "logs_policy" {
+  count = var.create_instance_profile && var.install_cloudwatch_agent ? 1 : 0
+  name  = "CloudWatchAgentPutLogsRetention"
+  role  = aws_iam_role.main[0].name
+
+  policy = jsonencode({
+    Version = "2012-10-17"
+    Statement = [
+      {
+        Action = [
+          "logs:PutRetentionPolicy",
+        ]
+        Effect   = "Allow"
+        Resource = "*"
+      },
+    ]
+  })
 }
 
 ############################
@@ -65,13 +130,29 @@ resource "aws_security_group" "main" {
   description = "ASG Group Security Group"
   vpc_id      = var.vpc_id
 
-  egress {
-    description = "All egress traffic"
-    from_port   = 0
-    to_port     = 0
-    protocol    = "-1"
-    cidr_blocks = ["0.0.0.0/0"]
+  dynamic "ingress" {
+    for_each = var.security_group_ingress
+    content {
+      description      = "Rule to allow port ${try(ingress.value.from_port, "")} inbound traffic"
+      from_port        = try(ingress.value.from_port, 22)
+      to_port          = try(ingress.value.to_port, 22)
+      protocol         = try(ingress.value.protocol, "tcp")
+      cidr_blocks      = try(ingress.value.cidr_blocks, ["0.0.0.0/0"])
+      ipv6_cidr_blocks = try(ingress.value.ipv6_cidr_blocks, ["::/0"])
+    }
   }
+
+  dynamic "egress" {
+    for_each = var.security_group_egress
+    content {
+      description = "Rule to allow outbound traffic"
+      from_port   = try(ingress.value.from_port, 0)
+      to_port     = try(ingress.value.to_port, 0)
+      protocol    = try(ingress.value.protocol, -1)
+      cidr_blocks = try(ingress.value.cidr_blocks, ["0.0.0.0/0"])
+    }
+  }
+
   tags = merge(
     {
       "Name"        = var.name
@@ -79,18 +160,6 @@ resource "aws_security_group" "main" {
     },
     var.other_tags,
   )
-}
-
-resource "aws_security_group_rule" "main" {
-  for_each                 = var.security_group_rules
-  security_group_id        = aws_security_group.main.id
-  description              = "Additional rules to add to the security group"
-  from_port                = lookup(each.value, "from_port", null)
-  to_port                  = lookup(each.value, "to_port", null)
-  protocol                 = lookup(each.value, "protocol", "tcp")
-  type                     = lookup(each.value, "type", null)
-  cidr_blocks              = lookup(each.value, "cidr_blocks", null)
-  source_security_group_id = lookup(each.value, "source_security_group_id", null)
 }
 
 ############################
@@ -246,8 +315,8 @@ resource "aws_launch_template" "main" {
   ebs_optimized                        = var.ebs_optimized
   image_id                             = var.image_id
   instance_type                        = var.instance_type
-  key_name                             = var.key_name
-  user_data                            = var.enable_monitoring ? base64encode(data.template_cloudinit_config.config.rendered) : var.user_data
+  key_name                             = var.create_key_pair ? aws_key_pair.main[0].key_name : var.key_name
+  user_data                            = var.install_cloudwatch_agent ? base64encode(data.template_cloudinit_config.config.rendered) : var.user_data
   vpc_security_group_ids               = length(var.network_interfaces) > 0 ? [] : concat(var.security_group_ids, [aws_security_group.main.id])
   default_version                      = var.default_version
   update_default_version               = var.update_default_version
@@ -377,6 +446,10 @@ resource "aws_launch_template" "main" {
     enabled = var.enable_monitoring
   }
 
+  iam_instance_profile {
+    name = var.create_instance_profile ? aws_iam_instance_profile.main[0].name : var.iam_instance_profile
+  }
+
   dynamic "network_interfaces" {
     for_each = var.network_interfaces
     content {
@@ -459,7 +532,7 @@ resource "aws_autoscaling_schedule" "main" {
 
 ########################################
 ## Autoscaling Policies Resources
-## Bellow are the resources to trigger autoscaling events and report to an email address (default)
+## Below are the resources to trigger autoscaling events and report to an email address (default)
 ########################################
 resource "aws_autoscaling_policy" "main" {
   for_each                  = var.autoscaling_policy