+++ versions.tf
@@ -7,6 +7,11 @@ terraform {
       version = ">= 4.0.0"
     }
 
+    null = {
+      source  = "hashicorp/null"
+      version = ">= 3.0.0"
+    }
+
     template = {
       source  = "hashicorp/template"
       version = ">= 2.0.0"