+++ CHANGELOG.md
@@ -5,12 +5,20 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-- Fix: Userdata partial success (some packages not installing)
 - Add: More options in the complete example
 - Feature: Ability for userdata packages to be installed in different OS flavours
 - Feature: Dashboards integration for Cloudwatch
 - Feature: Tag inheritance for created resources
 - Feature: Expand the complete example further
+- Feature: Add updated features from tf resource page
+
+## [1.2.0] - 2022-06-22
+### Changes
+- Fix: Userdata partial success (some packages not installing) specifically cloudwatch agent
+- Added required permissions for agent to send logs and metrics to cloudwatch
+- Added IAM instance profile feature for LC
+
+[1.2.0]: https://github.com/boldlink/terraform-aws-autoscaling/releases/tag/1.2.0
 
 ## [1.1.1] - 2022-06-10
 ### Changes